requirements.txt
vllm
fastapi
uvicorn
pyyaml
pydantic